requirements.txt
fastapi==0.114.1
pydantic==2.9.1
openai==0.28.0
uvicorn==0.30.6
starlette==0.38.5
aiohttp==3.10.5
async-timeout==4.0.3
httpx==0.27.2
python-dotenv==1.0.0
gunicorn==23.0.0
python-dotenv==1.0.1